requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.0.5

Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==5.0.0
bokeh
brotlipy==0.7.0

cffi==1.15.0
charset-normalizer==2.0.6
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.0
cycler==0.10.0
cytoolz==0.11.2

decorator==5.1.1
distributed==2.20.0
docstring-parser==0.7.3

fasteners==0.16.3
fonttools==4.25.0

h5py
HeapDict==1.0.1
hsluv==5.0.2

imagesize==1.2.0

importlib-resources==5.4.0
ipykernel==6.13.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
keyring==23.5.1
kiwisolver==1.3.1
locket==0.2.1
seaborn

MarkupSafe==2.1.2
matplotlib
matplotlib-inline==0.1.3
mkl-fft
mkl-random

msgpack==1.0.2
munkres==1.1.4
napari
#
napari-plugin-engine
napari-svg
nest-asyncio==1.5.4
networkx==2.6.3
numcodecs
numexpr
numpy
numpydoc==1.1.0

pandas
parso==0.8.3
partd==1.2.0
pickleshare==0.7.5
Pillow==9.0.1

prompt-toolkit==3.0.22


Pygments==2.11.2
PyOpenGL==3.1.6
pyOpenSSL==20.0.1
pyparsing==3.0.3
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyreadline==2.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2

PyWavelets
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==5.4.1

qtconsole==5.1.1
QtPy==2.0.0

requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0

scikit-learn
scipy

six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sphinxcontrib-applehelp==1.0.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tblib==1.7.0
tifffile==2022.5.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
twine==3.8.0
typing_extensions==4.0.1
urllib3==1.26.8
vispy==0.6.6
wcwidth==0.2.5
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.1
ypstruct==0.0.2
zarr==2.11.3
zict==2.0.0
zipp==3.5.0
